+++ api/feedback/index.php
@@ -5,6 +5,7 @@ require_once __DIR__ . '/../utils/auth_middleware.php';
 require_once __DIR__ . '/../utils/response.php';
 
 requireAuth();
+requireCsrf();
 
 $method = $_SERVER['REQUEST_METHOD'];