+++ api/stats/goals.php
@@ -4,6 +4,7 @@ require_once __DIR__ . '/../utils/auth_middleware.php';
 require_once __DIR__ . '/../utils/response.php';
 
 requireAuth();
+requireCsrf();
 
 $db = Database::getInstance()->getConnection();
 $userId = getCurrentUserId();